migrated the chart view
prepare for crop function.
