+++ monitor/viewer.html
@@ -174,6 +174,11 @@
           <div class="end-marker left"></div>
           <div class="end-marker right"></div>
         </div>
+        <div class="controls overlay_control">
+          <button class="overlay_button" id="btnScissor" aria-label="Crop" title="Crop">
+            <i data-feather="scissors"></i>
+          </button>
+        </div>
       </div>
       <div class="labels-layer">
         <span id="lblFloating" class="hidden">--:--:--.---</span>
